Use Chassis power state for power state checks
The current code was using the CurrentHostState property in all checks
relating to power being on, looking for a value of Running to mean power is
on.

That works fine for code that uses the x86-power-control repo to provide
CurrentHostState as the only possible values are Running and Off.
However the CurrentHostState provided by the phosphor-state-manager repo
has more than just Running and Off, such as DiagnosticMode or Quiesced.
The actual power may or may not be on in those latter two cases.

This commit changes the code to look at the CurrentPowerState property
on the chassis state object instead.  With x86-power-control, the
chassis state always matches the host state anyway, and with
phosphor-state-manager the CurrentPowerState property can only be On or
Off, just like with x86-power-control.  In both cases it accurately
reflects the main power state of the system.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I6343ea876c060f85d2c6eee0d25b3d7458885b72