Add framework for power state change callbacks
Create a PowerStateCallback class where other code can register
functions to call when the power state changes using the static
PowerStateCallback::addCallback() function.

The class keeps a vector of those functions and calls them from hooks in
the code that monitors power state changes.

Also, change to always create the power match so that it will be
available even when no sensors depend on it.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I10f66b79ff6c0074f934f0ac090c901f6bf5f136