hwmontemp: Create event logs for failed reads
When a sensor has hit enough consecutive read fails that it gets set to
nonfunctional with a sensor value of NaN, create an event log for it
with the I2C bus, address, sensor name, and other details included in
the event log.

Only create event logs when power is on, and only create one event log
per hardware device per power cycle.  i.e. if a temp sensor has two
temp*_input files and both fail, only create an event log for one of
them.

Only one event log is created per power cycle so there is no flooding of
event logs for flakey sensors that go in and out of readability.

When power is switched on, clear the history of fails, and look for any
sensors that are currently failed and create event logs for them.

Note the code doesn't use the I2C bus and address stored in the
I2CDevice object in the HwmonTempSensor object because it isn't created
for all device types.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I97ce0583ed9bc82b357b667fc631fa2c8b0d02a3