adcsensor: Match the sensor device name
Currently, ADCSensor daemon only accesses the sensor paths in
/sys/devices/platform/iio-hwmon or their device name is iio-hwmon. So
the daemon created the sensor path filter by comparing the output name
of /sys/class/hwmon/hwmon*/name with "iio_hwmon" string. But the linux
kernel doesn't limit the number of ADC devices by one. Such as AST2600
has two ADC devices. So the sensor paths are not only in
/sys/devices/platform/iio-hwmon but also in other paths such as
/sys/devices/platform/iio-hwmon-*. That why, the filter should not
exist.

Moreover, when there are some ADC devices, the sensor index of
different ADC devices can be dupplicated. So only matching the sensor
path index with the "Index" value in sensor configuration as the
current design is not enough. Therefore, ADCSensor daemon should add
more matching conditions. The device name in the device path can be
used as the addition checking.

This commit will remove the sensor path filter and supports using the
device name as the addition matching beside the sensor path index.
Because the current ADC sensor configurations do not have the device
name setting, the default name "iio-hwmon" will be used for backward
compatible. The "DevName" option will be add later in sensor
configuration file to support setting device name.

Below is an example of two sensor paths of two different devices with
the same index:
$readlink -f /sys/class/hwmon/hwmon11/in1_input
/sys/devices/platform/iio-hwmon/hwmon/hwmon11/in1_input
$readlink -f /sys/class/hwmon/hwmon10/in1_input
/sys/devices/platform/iio-hwmon-battery/hwmon/hwmon10/in1_input

Tested:
    1. Configures ADC sensors as below:
        { "Index": 0, "Name": "S0_0V8_VDDC_SOC", "Type": "ADC" }
    2. The configured sensor should be matched with the sensor path
        /sys/devices/platform/iio-hwmon/hwmon/hwmon11/in1_input.

Signed-off-by: ThuBaNguyen <thu@os.amperecomputing.com>
Change-Id: I32dbb582edc0e5c83e4ac63e7ae19e1158f4526f