hwmontemp: Support for devices on slot power (#12)
hwmontemp: Support for devices on slot power

There are a few models of IBM PCIe cards that have TMP435 temperature
sensors which are on a power domain controlled by the host and tracked
by the xyz.openbmc_project.State.Decorator.PowerState interface on the
corresponding PCIe slot D-Bus objects.  The devices won't have power
until some point after the host has been started, so the devices can't
even be detected until that point.

This commit adds support for that by creating a minimally invasive
SlotPowerManager class to watch for when the slots are turned on.  At
that point it will call the 'new_device' I2C driver interface to bind
the device driver so the hwmon instance is created, and call the
createSensors() function in HwmonTempMain.cpp so the appropriate sensor
D-Bus objects can then be created.

When the system is powered off, the driver will be left bound but the
HwmonTempSensor class will ask SlotPowerManager if the device is off so
it will just display NaN until the slot powers on again without
attempting any reads.  On subsequent boots the code will see the driver
has already been bound so it won't do it again.

To indicate to SlotPowerManager a device is on slot power it needs

"PowerStateOwner": "PCIeSlot",
"LocationCode": "<location code of slot>"

in the entity-manager D-Bus config.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: If88673f631de838a97884be1df7d5146c3c40d3e