adc: Only process CPUReq sensors when CPUs change
The adcsensor application watches for the Present property to change on
CPUs so that it can check if any ADC sensors that have the 'CPURequired'
configuration field need changes.

However, in this case it was also deleting and recreating non-CPU
related sensors, such as ones that measure the RTC battery voltage. This
is a problem because if that sensor is below threshold, it will behave
as if the threshold were newly asserted by emitting propertiesChanged
signals on the threshold alarm properties as well as emitting the
ThresholdAsserted signals again.

Fix this by changing the code to only process sensors that have the
CPURequired field when handling CPU presence changes.

Tested:
Check D-Bus signals when CPU presence changes versus when a threshold
configuration property changes and see the expected handling.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I8f00726e71c85e70203a5cb6ee5dd99b4fe9e9c1