hwmontemp: slot power: Check Bus/Addr validity
The code that handles temperature sensors that are on slot power pulls
the Bus and Address properties out of the D-bus config data.  If those
properties weren't valid, which could happen when a card is plugged in
an unsupported slot, it was crashing because they aren't uint64 values.

Check if those values are of type uint64_t before accessing them.  Also,
check if they even exist first, though they should if we got this far.

Signed-off-by: Matt Spinler <spinler@us.ibm.com>
Change-Id: I8b929c4669f4baf81faf87bad9047b909204730f