clang-tidy: reduce repo-specific flags
Adjust the clang-tidy settings so that it does not need to have
a bunch of values related to the libpeci subproject.  In general
we should be ignoring any header files from any subprojects since
they are not under direct control of the repository-under-test.

Change-Id: I64d4e4bae20ca27aed68081b4270fa0ec8fc0543
Signed-off-by: Patrick Williams <patrick@stwcx.xyz>
(cherry picked from commit d02ad49eb60dd0761686220812b5afabb862693e)